made clearFirestoreData throw error at invalid data
make clearFirestoreData throw error when projectId is not specfied.

+++ src/providers/firestore.ts
@@ -236,6 +236,9 @@ export type ClearFirestoreDataOptions = {
  */
 export function clearFirestoreData(options: ClearFirestoreDataOptions) {
   return new Promise((resolve, reject) => {
+    if (!options.projectId) {
+      throw new Error('projectId not specified');
+    }
     const config = {
       method: 'DELETE',
       hostname: 'localhost',
